[scanner] fix: resolve int64 type mismatch in server_ai_tokens.go
The getClaudeInfo() function was declaring token counter variables as 'int'
but s.tokens.GetUsage() returns 'int64' values. These were then assigned to
protocol.TokenCount fields which are int64, causing a type mismatch error.

Changed the variable declarations to int64 to match the return type and
the TokenCount field types.

+++ pkg/agent/server_ai_tokens.go
@@ -24,7 +24,7 @@ func (s *Server) getClaudeInfo() *protocol.ClaudeInfo {
 		providerNames = append(providerNames, p.DisplayName)
 	}
 
-	var sessionIn, sessionOut, todayIn, todayOut int
+	var sessionIn, sessionOut, todayIn, todayOut int64
 	if s.tokens != nil {
 		sessionIn, sessionOut, todayIn, todayOut = s.tokens.GetUsage()
 	}
